First pass at handling strings; some in/out choices may be backwards
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@151 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/libf/inq_dimlenf.c
@@ -21,6 +21,6 @@
 
 /* Prototypes for the Fortran interfaces */
 #include "mpifnetcdf.h"
-FORTRAN_API void FORT_CALL nfmpi_inq_dimlen_ ( int *v1, int *v2, size_t*v3, MPI_Fint *ierr ){
+FORTRAN_API void FORT_CALL nfmpi_inq_dimlen_ ( int *v1, int *v2, int *v3, MPI_Fint *ierr ){
     *ierr = ncmpi_inq_dimlen( *v1, *v2, v3 );
 }
